+++ grails-app/services/org/bbop/apollo/Gff3HandlerService.groovy
@@ -1,12 +1,9 @@
 package org.bbop.apollo
 
 import org.apache.commons.lang.WordUtils
-import org.bbop.apollo.geneProduct.GeneProduct
-import org.bbop.apollo.go.GoAnnotation
 import org.bbop.apollo.gwt.shared.FeatureStringEnum
 import org.bbop.apollo.sequence.Strand
 import org.grails.plugins.metrics.groovy.Timed
-import org.springframework.format.datetime.DateFormatter
 import java.text.SimpleDateFormat
 
 
@@ -53,6 +50,9 @@ class Gff3HandlerService {
         writeObject.attributesToExport.add(FeatureStringEnum.COMMENTS.value);
         writeObject.attributesToExport.add(FeatureStringEnum.DATE_CREATION.value);
         writeObject.attributesToExport.add(FeatureStringEnum.DATE_LAST_MODIFIED.value);
+        writeObject.attributesToExport.add(FeatureStringEnum.IS_FMIN_PARTIAL.value);
+        writeObject.attributesToExport.add(FeatureStringEnum.IS_FMAX_PARTIAL.value);
+        writeObject.attributesToExport.add(FeatureStringEnum.OBSOLETE.value);
 
         if (!writeObject.file.canWrite()) {
             throw new IOException("Cannot write GFF3 to: " + writeObject.file.getAbsolutePath());
@@ -354,6 +354,15 @@ class Gff3HandlerService {
             if (writeObject.attributesToExport.contains(FeatureStringEnum.GENE_PRODUCT.value) && feature.geneProducts) {
                 String productString  = geneProductService.convertGeneProductsToGff3String(feature.geneProducts)
                 attributes.put(FeatureStringEnum.GENE_PRODUCT.value, encodeString(productString))
+            }
+            if (writeObject.attributesToExport.contains(FeatureStringEnum.IS_FMIN_PARTIAL.value) &&  feature.featureLocation.isFminPartial) {
+                attributes.put(FeatureStringEnum.IS_FMIN_PARTIAL.value, feature.featureLocation.isFminPartial.toString())
+            }
+            if (writeObject.attributesToExport.contains(FeatureStringEnum.OBSOLETE.value) &&  feature.isObsolete) {
+                attributes.put(FeatureStringEnum.OBSOLETE.value, feature.isObsolete.toString())
+            }
+            if (writeObject.attributesToExport.contains(FeatureStringEnum.IS_FMAX_PARTIAL.value) &&  feature.featureLocation.isFmaxPartial) {
+                attributes.put(FeatureStringEnum.IS_FMAX_PARTIAL.value, feature.featureLocation.isFmaxPartial.toString())
             }
             if (writeObject.attributesToExport.contains(FeatureStringEnum.STATUS.value) && feature.getStatus() != null) {
                 attributes.put(FeatureStringEnum.STATUS.value, encodeString(feature.getStatus().value));