+++ src/gwt/org/bbop/apollo/gwt/client/GeneDetailPanel.java
@@ -22,7 +22,6 @@ import org.bbop.apollo.gwt.client.rest.AnnotationRestService;
 import org.bbop.apollo.gwt.client.rest.AvailableStatusRestService;
 import org.bbop.apollo.gwt.client.rest.RestService;
 import org.bbop.apollo.gwt.shared.FeatureStringEnum;
-import org.bbop.apollo.gwt.shared.PermissionEnum;
 import org.gwtbootstrap3.client.ui.*;
 import org.gwtbootstrap3.extras.bootbox.client.Bootbox;
 import org.gwtbootstrap3.extras.bootbox.client.callback.ConfirmCallback;
@@ -41,7 +40,7 @@ public class GeneDetailPanel extends Composite {
     interface AnnotationDetailPanelUiBinder extends UiBinder<Widget, GeneDetailPanel> {
     }
 
-    private static AnnotationDetailPanelUiBinder ourUiBinder = GWT.create(AnnotationDetailPanelUiBinder.class);
+    private static final AnnotationDetailPanelUiBinder ourUiBinder = GWT.create(AnnotationDetailPanelUiBinder.class);
     @UiField(provided = true)
     SuggestBox nameField;
     @UiField
@@ -74,6 +73,12 @@ public class GeneDetailPanel extends Composite {
     Button gotoAnnotation;
     @UiField
     Button annotationIdButton;
+    @UiField
+    InlineCheckBox partialMin;
+    @UiField
+    InlineCheckBox partialMax;
+    @UiField
+    InlineCheckBox obsoleteButton;
 
     private SuggestedNameOracle suggestedNameOracle = new SuggestedNameOracle();
 
@@ -100,6 +105,12 @@ public class GeneDetailPanel extends Composite {
     void handleSymbolChange(ChangeEvent e) {
         String updatedName = symbolField.getText();
         internalAnnotationInfo.setSymbol(updatedName);
+        updateGene();
+    }
+
+    @UiHandler("obsoleteButton")
+    void handleObsoleteChange(ChangeEvent e) {
+        internalAnnotationInfo.setObsolete(obsoleteButton.getValue());
         updateGene();
     }
 
@@ -175,6 +186,15 @@ public class GeneDetailPanel extends Composite {
         updateGene();
     }
 
+
+    @UiHandler({"partialMin", "partialMax"})
+    void handlePartial(ChangeEvent e){
+        internalAnnotationInfo.setPartialMin(partialMin.getValue());
+        internalAnnotationInfo.setPartialMax(partialMax.getValue());
+        updatePartials();
+    }
+
+
     private void checkSyncButton(){
         Set<AnnotationInfo> childAnnotations = internalAnnotationInfo.getChildAnnotations();
         if(childAnnotations.size()==1){
@@ -193,6 +213,8 @@ public class GeneDetailPanel extends Composite {
         descriptionField.setEnabled(editable);
         synonymsField.setEnabled(editable);
         deleteAnnotation.setEnabled(editable);
+        partialMin.setEnabled(editable);
+        partialMax.setEnabled(editable);
 
         if(!editable || this.internalAnnotationInfo==null){
             syncNameButton.setEnabled(false);
@@ -202,6 +224,29 @@ public class GeneDetailPanel extends Composite {
         }
     }
 
+    private void updatePartials() {
+        setEditable(false);
+        RequestCallback requestCallback = new RequestCallback() {
+            @Override
+            public void onResponseReceived(Request request, Response response) {
+                JSONValue returnValue = JSONParser.parseStrict(response.getText());
+                setEditable(true);
+                MainPanel.annotatorPanel.setSelectedChildUniqueName(null);
+            }
+
+            @Override
+            public void onError(Request request, Throwable exception) {
+                Bootbox.alert("Error updating gene: " + exception);
+                setEditable(true);
+            }
+        };
+//        RestService.sendRequest(requestCallback, "annotator/updateFeature/", AnnotationRestService.convertAnnotationInfoToJSONObject(this.internalAnnotationInfo));
+        JSONObject data = AnnotationRestService.convertAnnotationInfoToJSONObject(this.internalAnnotationInfo);
+        data.put(FeatureStringEnum.ORGANISM.getValue(),new JSONString(MainPanel.getInstance().getCurrentOrganism().getId()));
+
+        RestService.sendRequest(requestCallback, "annotator/updatePartials/", data);
+
+    }
 
     private void updateGene() {
         setEditable(false);
@@ -236,6 +281,9 @@ public class GeneDetailPanel extends Composite {
         suggestedNameOracle.setOrganismName(MainPanel.getInstance().getCurrentOrganism().getName());
         suggestedNameOracle.setFeatureType("sequence:" + annotationInfo.getType());
         nameField.setText(internalAnnotationInfo.getName());
+        partialMin.setValue(internalAnnotationInfo.getPartialMin());
+        partialMax.setValue(internalAnnotationInfo.getPartialMax());
+        obsoleteButton.setValue(internalAnnotationInfo.getObsolete());
         symbolField.setText(internalAnnotationInfo.getSymbol());
         typeField.setText(internalAnnotationInfo.getType());
         synonymsField.setText(internalAnnotationInfo.getSynonyms());